admin books and categories completed, updated database

+++ admin/templates/view_all_books.php
@@ -48,7 +48,7 @@
 								<td class="text-center"><?php echo $originalprice ?></td>
 								<td class="text-center"><a href="view_more_books.php?bookid=<?php echo $bookid ?>" class="btn btn-primary">View More</a></td>
 								<td class="text-center"><a href="books.php?source=edit_book&editId=<?php echo $bookid ?>" class="btn btn-primary">Edit</a></td>
-								<td class="text-center"><a href="books.php?delete=<?php echo $bookid ?>" class="btn btn-danger">Delete</a></td>
+								<td class="text-center"><a onClick="javascript: return confirm('Are you sure you want to delete this review'); " href="books.php?delete=<?php echo $bookid ?>" class="btn btn-danger">Delete</a></td>
 							</tr>
 				<?php
 						}
