+++ admin/templates/view_all_users.php
@@ -46,7 +46,7 @@
 								<td class="text-center"><?php echo $contact ?></td>
 								<td class="text-center"><?php echo $user_category ?></td>
 								<td class="text-center"><a href="users.php?source=edit_user&edit=<?php echo $username ?>" class="btn btn-primary">Edit</a></td>
-								<td class="text-center"><a href="users.php?delete=<?php echo $username ?>" class="btn btn-danger">Delete</a></td>
+								<td class="text-center"><a onClick="javascript: return confirm('Are you sure you want to delete this review'); " href="users.php?delete=<?php echo $username ?>" class="btn btn-danger">Delete</a></td>
 							</tr>
 				<?php
 						}
